Without Vo&Converter confoguration

+++ src/main/java/com/onda/personnel/service/impl/EmployeeServiceImpl.java
@@ -8,6 +8,7 @@ package com.onda.personnel.service.impl;
 import com.onda.personnel.bean.Employee;
 import com.onda.personnel.dao.EmployeeDao;
 import com.onda.personnel.service.EmployeeService;
+import com.onda.personnel.service.WorkService;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -21,6 +22,9 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     @Autowired
     EmployeeDao employeeDao;
+
+    @Autowired
+    private WorkService workService;
 
     @Override
     public Employee findByMatricule(Integer matricule) {
@@ -58,18 +62,20 @@ public class EmployeeServiceImpl implements EmployeeService {
         Employee checkEmployee = findByMatricule(matricule);
         if (checkEmployee == null) {
             return -1;
-
         } else {
-            employeeDao.delete(checkEmployee);
-            return 1;
+            if (-1 == -1) {
+                return -2;
+            } else {
+                employeeDao.delete(checkEmployee);
+                return 1;
+            }
         }
     }
-    
+
     @Override
     public List<Employee> findAll() {
-return employeeDao.findAll();
+        return employeeDao.findAll();
     }
-
 
     public EmployeeDao getEmployeeDao() {
         return employeeDao;
@@ -79,5 +85,4 @@ return employeeDao.findAll();
         this.employeeDao = employeeDao;
     }
 
-    
 }
